pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tmux-orchestrator"
version = "0.1.0"
description = "Orchestrate AI agents (Claude Code and custom) inside tmux panes"
requires-python = ">=3.11"
dependencies = [
    "libtmux>=0.28",
    "textual>=0.60",
    "fastapi>=0.110",
    "uvicorn[standard]",
    "pyyaml>=6",
    "typer>=0.12",
    "rich>=13",
    "websockets>=12",
]

[project.scripts]
tmux-orchestrator = "tmux_orchestrator.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tmux_orchestrator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
